Tweaks To Timeline
