+++ src/app/page.js
@@ -1,11 +1,72 @@
 "use client";
 import Player from "@/components/Player";
 import { useState } from "react";
-import Timeline from "./omprasad/timeline";
+import Timeline from "@/components/Timeline/timeline";
+
 export default function Home() {
+  const [counter, setCounter] = useState(0);
+
+  const headingClicked = () => {
+    setCounter(counter + 1);
+  };
+  if (counter == 3) {
+    document.getElementById("rocket").classList.add("animate-spin");
+  }
+
+  const handleStop = () => {
+    setCounter(0);
+    document.getElementById("rocket").classList.remove(["animate-spin"]);
+  };
+
   return (
     <>
-    <Timeline/>
+      <main className="flex min-h-screen min-w-screen flex-col items-center justify-center p-12 xl:p-0 md:text-2xl bg-blue-100">
+        <div className="drop-shadow-xl flex-col flex gap-10 items-center ">
+          <Player />
+          <div>
+            <h1
+              className="text-center font-black text-3xl md:text-5xl break-keep break-words mb-4 cursor-pointer"
+              onClick={headingClicked}
+            >
+              HackOdisha 3.0
+            </h1>
+            <h2 className="text-center font-black text-xl md:text-3xl break-keep break-words">
+              Coming Soon{" "}
+              <span className="group">
+                <p
+                  id="rocket"
+                  className="cursor-pointer inline-block"
+                  onClick={handleStop}
+                >
+                  🚀
+                </p>
+                <span className="absolute opacity-0 group-hover:opacity-100 text-xs duration-100 ease-in   ">
+                  {counter < 3
+                    ? "Tap HackOdisha Three times to see me rotate"
+                    : "Tap Me to Stop"}
+                </span>
+              </span>
+            </h2>
+          </div>
+          <a
+            href="https://twitter.com/hackodisha"
+            target="_blank"
+            rel="noreferrer"
+          >
+            <button className="px-4 py-2 bg-blue-200 w-fit hover:bg-blue-300 ease-in duration-100 rounded-lg ">
+              <span className="inline-flex items-center gap-4">
+                Check Status
+                <svg viewBox="0 0 24 24" aria-hidden="true" className="w-4 h-4">
+                  <g>
+                    <path d="M23.643 4.937c-.835.37-1.732.62-2.675.733.962-.576 1.7-1.49 2.048-2.578-.9.534-1.897.922-2.958 1.13-.85-.904-2.06-1.47-3.4-1.47-2.572 0-4.658 2.086-4.658 4.66 0 .364.042.718.12 1.06-3.873-.195-7.304-2.05-9.602-4.868-.4.69-.63 1.49-.63 2.342 0 1.616.823 3.043 2.072 3.878-.764-.025-1.482-.234-2.11-.583v.06c0 2.257 1.605 4.14 3.737 4.568-.392.106-.803.162-1.227.162-.3 0-.593-.028-.877-.082.593 1.85 2.313 3.198 4.352 3.234-1.595 1.25-3.604 1.995-5.786 1.995-.376 0-.747-.022-1.112-.065 2.062 1.323 4.51 2.093 7.14 2.093 8.57 0 13.255-7.098 13.255-13.254 0-.2-.005-.402-.014-.602.91-.658 1.7-1.477 2.323-2.41z"></path>
+                  </g>
+                </svg>
+              </span>
+            </button>
+          </a>
+        </div>
+      </main>
+      <Timeline/>
     </>
   );
 }
