+++ src/components/Timeline/timeline.css
@@ -17,7 +17,7 @@
 	height: auto;
   }
   body{
-	background-image: url(./background.png);
+	background-image: url(./assets/background.png);
 	background-position: center;
 	background-repeat: no-repeat;
 	background-size: cover;
@@ -35,8 +35,9 @@
      padding: 5px;
 	}
 	.timeline{
-		background-image: url(./Vector\ 1.svg);
+		background-image: url(./assets/Vector\ 1.svg);
 		background-position: center;
+		background-size: contain;
 		background-repeat: no-repeat;
 		margin-top: 2rem;
 	 }
@@ -77,11 +78,11 @@
 	  letter-spacing: -0.5px;
 	}
 	.contentHead{
-	  width:103px;
+	  width:fit-content;
 	}
 	.contentHead h3 {
 	  color: #CA40CD;
-	  width:103px;
+	  width:fit-content;
 	  font-family: Syoog;
 	  font-size: 15px;
 	  font-style: normal;
@@ -95,9 +96,10 @@
   @media only screen and (min-width:550px) and (max-width:750px) {
 	
 	.timeline{
-		background-image: url(./Vector\ 2.svg);
+		background-image: url(./assets/Vector\ 2.svg);
 		background-position: center;
 		background-repeat: no-repeat;
+		background-size: contain;
 		margin-top: 3rem;
 	 }
     .img1{
@@ -137,11 +139,11 @@
 	  letter-spacing: -0.5px;
 	}
 	.contentHead{
-	  width:103px;
+	  width:fit-content;
 	}
 	.contentHead h3 {
 	  color: #CA40CD;
-	  width:103px;
+	  width:fit-content;
 	  font-family: Syoog;
 	  font-size: 15px;
 	  font-style: normal;
@@ -153,7 +155,7 @@
   
   @media only screen and (min-width:750px) and (max-width:1000px) {
 	.timeline{
-		background-image: url(./Vector\ 3.svg);
+		background-image: url(./assets/Vector\ 3.svg);
 		background-position: center;
 		background-repeat: no-repeat;
 		margin-top: 3rem;
@@ -195,11 +197,11 @@
 	  letter-spacing: -0.5px;
 	}
 	.contentHead{
-	  width:103px;
+	  width:fit-content;
 	}
 	.contentHead h3 {
 	  color: #CA40CD;
-	  width: 103px;
+	  width: fit-content;
 	  font-family: Syoog;
 	  font-size: 15px;
 	  font-style: normal;
@@ -209,7 +211,7 @@
   }
   @media only screen and (min-width:1000px) and (max-width:1300px) {
 	.timeline{
-		background-image: url(./Vector\ 4.svg);
+		background-image: url(./assets/Vector\ 4.svg);
 		background-position: center;
 		background-repeat: no-repeat;
 		margin-top: 4rem;
@@ -251,11 +253,11 @@
 	  letter-spacing: -0.5px;
 	}
 	.contentHead{
-	  width:103px;
+	  width:fit-content;
 	}
 	.contentHead h3 {
 	  color: #CA40CD;
-	  width: 103px;
+	  width: fit-content;
 	  font-family: Syoog;
 	  font-size: 15px;
 	  font-style: normal;
@@ -266,10 +268,11 @@
   
   @media only screen and (min-width:1300px) {
 	.timeline{
-		background-image: url(./Vector\ 5.svg);
+		background-image: url(./assets/Vector\ 5.svg);
 		background-position: center;
 		background-repeat: no-repeat;
-		margin-top:5rem;
+		margin-top:2rem;
+		padding-top: 4rem;
 	 }
 	.img1{
 		width:300px;
@@ -304,7 +307,7 @@
 	  width: 235px;
 	}
 	.contentHead h3 {
-	  width: 235px;
+	  width: fit-content;
 	  color: #ca40cd;
 	  font-family: Syoog;
 	  font-size: 35px;
